Disable testing with asan for now
Asan was failing with

```
==2022==ERROR: AddressSanitizer: use-after-poison on address 0x7f6ff0370010 at pc 0x7f7014306aec bp 0x7fffc216d170 sp 0x7fffc216d168
READ of size 8 at 0x7f6ff0370010 thread T0
    #0 0x7f7014306aeb in RB_BUILTIN_TYPE /home/runner/.rubies/ruby-asan/include/ruby-3.5.0+2/ruby/internal/value_type.h:191:30
    #1 0x7f7014306aeb in valid_method_entry /home/runner/work/direct-bind/direct-bind/tmp/x86_64-linux/direct_bind_native_extension/3.5.0/../../../../ext/direct_bind_native_extension/direct_bind.c:142:51
    #2 0x7f7014306aeb in find_cfunc /home/runner/work/direct-bind/direct-bind/tmp/x86_64-linux/direct_bind_native_extension/3.5.0/../../../../ext/direct_bind_native_extension/direct_bind.c:113:10
```

now I see in the Ruby sources that in similar situations (in iseq.c)
the VM is doing dark magic with "unpoison" and "poison" of objects
as it iterates the objspace.

But... I don't know enough about asan to understand yet if that's
correct, and I don't like the idea of cargo culting asan stuff without
understanding it "just to shut it up", which seems to somewhat
defeat the objective.

So... let's come back to this later.